Add a WebAppProvider D-Bus interface
This commit adds a D-Bus service that exposes the web apps managed by
Epiphany so that a client can enumerate them, install a new one, or
remove an installed one. The intended client is GNOME Software, since I
am soon going to add back a webapp plugin to Software which will make
use of this interface. This is part of a larger effort to improve the
support for Progressive Web Apps in GNOME (though Epiphany's web apps do
not support PWA manifest features).

The great thing about having this be a service exposed by Epiphany,
rather than having Software try to install Epiphany-compatible web apps
on its own as it used to do, is that the implementation of how to create
and manage the web apps can stay in Epiphany, and there can never be
disagreement between Epiphany and Software about the proper on-disk
format for them (e.g. the algorithm used for generating the app ID from
the name).

The goal for the PWA project is to support Flatpak'd Epiphany, whereas
currently only non-Flatpak Epiphany can do web apps. This will be
accomplished with new portals to allow installing and removing the
.desktop launchers. The Flatpak support requirement is reflected in the
design of the API here, specifically the install_token parameter for the
InstallWebApp method. This token is to be acquired by the client of the
D-Bus interface (Software) so that the installation can be achieved
without any additional user interaction since the user would've already
clicked "Install" in Software. Web app installation directly via
Flatpak'd Epiphany's UI would involve a dialog created by the portal;
this is because we don't sandboxed applications in general to be able to
create desktop launchers without user interaction.

The UninstallWebApp method by contrast doesn't require such a token,
because the portal can ensure that only apps created by an application
are deleted by that application.